CXP-1215: move Settings in CatalogEdit components
