+++ components/catalogs/front/src/components/CatalogEdit/components/Settings.test.tsx
