+++ components/catalogs/front/src/components/CatalogEdit/components/__mocks__/Settings.tsx
