--- components/catalogs/front/src/components/Settings/index.ts
@@ -1,2 +1,0 @@
-/* istanbul ignore file */
-export {Settings} from './Settings';
